added a method1 and modified the code partially suitable for it

+++ src/GraphStructure/Node.java
@@ -7,6 +7,6 @@ public interface Node
 public Iterator<Connection> listConnections();
 public void addConnection(String conname,Connection c);
 public void removeConnection(String conname);
-public void connect(Node n,Double weight);
+public void connect(Node n,int weight);
 public void disconnect(Node n);
 }
